Added a new folder to keep the repos in
This way, we don't have to rely on github ar anything like that for
testing
purposes. We can just populate the folder with repos, point the tests at
them and let them crunch. This should prove really usefull in the later
part
of our project.

For now this only contains a simple linear repo with three commits.
Basic
but it should be enough for what we need now.

Feel free to add some more :)