Rewrote the CSVCommitPrinter
I rewrote the printer to take into account the new order of things. The
result a lot cleaner and makes the whole thing 100% decoupled from
everything else.

Of course, I had to destroy the tests and a lot of things around to make
this happen. But the end result will be a lot cleaner and nicer. So I
think
that it is worth it.