Added a check to the PipelineFilter
The check makes sure we only add DataCollectors to the collectors list.

It's a hack to get round the single inheritance problem in java.
Because I
do not want to modify the gitective core (yet), I resorted to this so
that I
can guarantee some kind of type checking. It's at runtime but it's
better
than nothing.