Changed the CSVCommitPrinter
It now takes the whole PipelineFilter as argument in the constructor,
instead
of the collectors. This is needed so that any changes to the filter in
subclasses
will be propagated to the final aggregator.

The aggregator gets the list only when starting the analysis, therefore
any
changes will be picked up.