Added a test for the delete filter
I needed to check some the check the assumption that simply deleting
contents from a file will not mark the whole diff as a delete.

As I found out, deleting the contents marks it as a MODIFY, so we are
safe.

The test is a good sanity check, so I think that it's worth for it stay.