Parameterized the NonSourceFileBlacklister
Since we are using it for a number of languages, it makes
sense to parameterize it so it's easily reusable. Now, you
have to give it a list of extension you consider to be source
files.

To make life easier I created a few predefined lists for Java
source files and "all" (for now Java, C, C++ and Python).
More can be added if need may be.